refactor(calculator): 값 객체의 이름을 StringValue에서 PositiveInteger로 변경
- 좀 더 명확한 의미를 갖는 객체이름으로 변경
