+++ src/main/java/calculator/PositiveInteger.java
@@ -1,0 +1,18 @@
+package calculator;
+
+public class PositiveInteger {
+    private final String value;
+
+    public PositiveInteger(String value, PositiveIntegerValidator stringValidator) {
+        stringValidator.validation(value);
+        this.value = value;
+    }
+
+    public static PositiveInteger of(String value) {
+        return new PositiveInteger(value, new DefaultPositiveIntegerValidator());
+    }
+
+    public int toInt() {
+        return Integer.parseInt(value);
+    }
+}
